requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Zoom API
pyjwt==2.8.0

# YouTube API
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Google Sheets & Calendar
gspread==6.0.2

# Notion API
notion-client==2.2.1

# OpenAI (Whisper & GPT)
openai==1.12.0

# Audio Processing
pydub==0.25.1

# Environment Variables
python-dotenv==1.0.1

# Validation
pydantic==2.6.0
pydantic-settings==2.1.0

# Background Tasks
celery==5.3.6
redis==5.0.1

# Utilities
tenacity==8.2.3
python-dateutil==2.8.2

# Slack Notifications
slack-sdk==3.27.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
